🤖 Refactor : 코드 리팩토링
1. Lotto와 관련된 예외를 모아서 model 패키지로 이동
2. Money에 존재하는 예외 메시지는 Lotto와 관련있지 않고 Money Dto 그 자체적인 정의기 때문에 따로 빼두지 않음
