+++ src/main/java/lotto/model/dto/Money.java
@@ -11,6 +11,7 @@ public class Money {
     }
     private void validateMoney(int money){
         if(money < 0 )
+            // Lotto와는 관련 없는 에러 메시지기 때문에 따로 enum 처리하지 않음
             throw new IllegalArgumentException("[ERROR] 구입 금액은 0 이상일 때 입력할 수 있어요.");
     }
 }
